refactor: extract dashboard logic to DashboardService
Refs: #1
